Create pages for all pages in the Login folder
