+++ src/pages/Register/index.scss
@@ -1,0 +1,41 @@
+#register {
+    max-width: 400px;
+    margin: auto;
+    .container {
+        background: linear-gradient(359.93deg, #FFFFFF 10.95%, #FCFBFF 121.59%);
+        border: 1px solid #EEEEEE;
+        box-shadow: 0px 23px 35px -12px rgba(31, 6, 87, 0.08);
+        border-radius: 10px;
+        padding: 32px 16px 24px;
+        margin-top: 36px;
+        .page-title {
+            @media screen and (max-width: 768px) {
+                margin-bottom: 1rem !important;
+            }
+        }
+        .dropdown {
+            .dropdown-toggle {
+                color: #0a060e;
+                background-color: white!important;
+                display: flex;
+                justify-content: space-between;
+                .fa {
+                    font-size: 26px;
+                }
+            }
+            .dropdown-menu {
+                a {
+                    color: #0a060e;
+                }
+                &.show {
+                    height: 300px;
+                    overflow: auto;
+                }
+            }
+        }
+    }
+    @media screen and (max-width: 768px) {
+        height: auto !important;
+        margin-top: 4rem !important;
+    }
+}
