+++ src/pages/VerifyEmail/index.jsx
