+++ src/pages/VerifyEmail/index.scss
@@ -1,0 +1,28 @@
+#verifyEmail {
+    max-width: 400px;
+    margin: auto;
+    .container {
+        background: linear-gradient(359.93deg, #FFFFFF 10.95%, #FCFBFF 121.59%);
+        border: 1px solid #EEEEEE;
+        box-shadow: 0px 23px 35px -12px rgba(31, 6, 87, 0.08);
+        border-radius: 10px;
+        padding: 32px 16px 24px;
+        .btn {
+            &.bg-blue {
+                background: #6309FF;
+                &:focus {
+                    opacity: 0.8;
+                }
+            }
+            &.bg-white {
+                color: #6309FF;
+                border: 2px solid #6309FF;
+                &:focus {
+                    background: #6309FF!important;
+                    color: #fff;
+                    opacity: 0.8;
+                }
+            }
+        }
+    }
+}
